Give the UI, logging, highlight and cursor-shape enums single homes
Four more families the transpiler copied per module:

- `kUI*` (35 files) and `kLineFlag*` join their `UIExtension`/`LineFlags`
  aliases in types/ui.rs, matching the `kObjectType*` precedent.
- `LOGLVL_*` (39 files) settles in log.rs. Its `LOGLVL_INF` had no home
  at all — log.rs declared the other three — so the file grows by the
  one constant and its comment; `just refresh --allow-growth` records
  that, against 39 files losing their copies.
- The highlight attribute bits (`HL_BOLD` … `HL_UNDERLINE_MASK`) settle
  in highlight.rs. As with the mode masks they were `c_uint` while
  `rgb_ae_attr`/`cterm_ae_attr` are `int32_t`, so 221 uses carried a
  cast; typing them to match retires those. `C2Rust_Unnamed_14` becomes
  `HlAttrFlags`. The syntax flags that share the `HL_` prefix are a
  different enum and stay in syntax.rs.
- `SHAPE_*` settles in cursor_shape.rs, with the same `c_int` retyping
  for the 35 index casts. `C2Rust_Unnamed_0` becomes `ShapeIdx`.

+++ crates/nvim/src/nvim/terminal.rs
@@ -642,9 +642,9 @@ unsafe fn get_rgb(state: *mut VTermState, mut color: VTermColor) -> c_int {
 fn get_underline_hl_flag(attrs: VTermScreenCellAttrs) -> c_int {
     match attrs.underline() {
         0 => 0,
-        2 => HL_UNDERDOUBLE as c_int,
-        3 => HL_UNDERCURL as c_int,
-        _ => HL_UNDERLINE as c_int,
+        2 => HL_UNDERDOUBLE,
+        3 => HL_UNDERCURL,
+        _ => HL_UNDERLINE,
     }
 }
